Add StatusUpdate function in a seperate file
